Break lineFunctions into separate files
